+++ libteol0/teonet_l0_client.h
@@ -296,9 +296,9 @@ extern "C" {
 TEOCLI_API void teoLNullInit();
 TEOCLI_API void teoLNullCleanup();
 
-TEOCLI_API teoLNullConnectData *teoLNullConnect(const char *server, int port, 
+TEOCLI_API teoLNullConnectData *teoLNullConnect(const char *server, int16_t port,
     PROTOCOL connection_flag);
-TEOCLI_API teoLNullConnectData* teoLNullConnectE(const char *server, int port,
+TEOCLI_API teoLNullConnectData* teoLNullConnectE(const char *server, int16_t port,
         teoLNullEventsCb event_cb, void *user_data, PROTOCOL connection_flag);
 TEOCLI_API void teoLNullDisconnect(teoLNullConnectData *con);
 TEOCLI_API void teoLNullShutdown(teoLNullConnectData *con);
